Add Login, Logout, & Registration Pages.
Add Routes for Registering, Logging In, & Logging Out.

Remembering logins is done by saving the Auth Token & User ID in
LocalStorage. When the application is reloaded, the Token & ID are used
to reauthenticate the user against the backend.

The `Add Routine` page is hidden for Anonymous users & the `Add
Exercise` page & Edit/Delete Exercise buttons are hidden for
non-Administrators.