Send Authentication Token with API Requests
The backend now requires the auth token for Create requests, so we send
the Authentication Token via the `Auth-Token` HTTP Header when making
requests, if the Authentication Status is `Authorized`.

We do this for all types of requests, since the backend will eventually
require this for Read, Update, Delete, & Lists requests as well.